Refactor how program tests are run
